pyproject.toml
[tool.poetry]
name = "beach-fractal"
version = "0.1.0"
description = "Generator for the fractal pattern people form when picking a spot on the beach."
authors = ["malyvsen <5940672+malyvsen@users.noreply.github.com>"]
license = "GPLv3"

[tool.poetry.dependencies]
python = "^3.9"
streamlit = "^0.83.0"
Pillow = "^8.2.0"
numpy = "^1.21.0"
stqdm = "^0.0.4"

[tool.poetry.dev-dependencies]
black = {version = "^21.6b0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
